libf

+++ libft.a/ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: dravaono <dravaono@student42nice.fr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/31 14:54:03 by dravaono          #+#    #+#             */
-/*   Updated: 2023/03/31 15:01:06 by dravaono         ###   ########.fr       */
+/*   Updated: 2023/04/04 15:45:19 by dravaono         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strchr(const char *s, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
