+++ libft.a/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: dravaono <dravaono@student42nice.fr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/27 16:10:48 by dravaono          #+#    #+#             */
-/*   Updated: 2023/03/31 15:32:44 by dravaono         ###   ########.fr       */
+/*   Updated: 2023/04/03 16:15:45 by dravaono         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
